+++ src/0_generalSettings.h
@@ -22,12 +22,12 @@
 //#define CORE_DEBUG // Don't use this!
 
 // EEPROM settings ---------------------------------------------------------------------------------------------------------------------------------
-uint8_t eeprom_id = 2; // change this id (between 1 and 255, compare with serial monitor), if you want to restore EEPROM defaults (only executed 1x) <<------------- NOTE!
+uint8_t eeprom_id = 5; // change this id (between 1 and 255, compare with serial monitor), if you want to restore EEPROM defaults (executed if different) <<------------- NOTE!
 //#define ERASE_EEPROM_ON_BOOT // EEPROM is completely overwritten, if defined! Never define it, vehicle will not work!
 // only define it in order to clean up junk from old projects in your EEPROM
 
 // Wireless settings -------------------------------------------------------------------------------------------------------------------------------
-#define ENABLE_WIRELESS // Define this, if you want to use an ESP-Now wireless trailer or the WiFi configuration via 192.168.4.1
+//#define ENABLE_WIRELESS // Define this, if you want to use an ESP-Now wireless trailer or the WiFi configuration via 192.168.4.1
 
 /* Wifi & ESP-Now and ESP-Now transmission power: less power = less speaker noise & longer battery life. Valid options are:
 WIFI_POWER_19_5dBm = 78     // full power